Data Viewer command registration into extension-side
