+++ src/webviews/extension-side/serviceRegistry.web.ts
@@ -13,6 +13,7 @@ import { VariableViewProvider } from './variablesView/variableViewProvider';
 import { JupyterVariableDataProvider } from './dataviewer/jupyterVariableDataProvider';
 import { JupyterVariableDataProviderFactory } from './dataviewer/jupyterVariableDataProviderFactory';
 import { IJupyterVariableDataProvider, IJupyterVariableDataProviderFactory } from './dataviewer/types';
+import { DataViewerCommandRegistry } from './dataviewer/dataViewerCommandRegistry';
 
 export function registerTypes(serviceManager: IServiceManager, _isDevMode: boolean) {
     serviceManager.add<IWebviewViewProvider>(IWebviewViewProvider, WebviewViewProvider);
@@ -26,5 +27,9 @@ export function registerTypes(serviceManager: IServiceManager, _isDevMode: boole
     serviceManager.addSingleton<IJupyterVariableDataProviderFactory>(
         IJupyterVariableDataProviderFactory,
         JupyterVariableDataProviderFactory
+    );
+    serviceManager.addSingleton<IExtensionSingleActivationService>(
+        IExtensionSingleActivationService,
+        DataViewerCommandRegistry
     );
 }
